Heat Map with Time using Folium
The Heat Map with Time is an interactive visualization that shows concentration of data over the time and, different from normal heat maps, this kind does not present relation between variables. It could also be an alternative for bar or line charts where it is wanted to express the gradual changes over the time.

In this case I am using the NYC traffic accidents dataset in order to show in which time of the day there is higher occurrence of incidents. The dataset can be found at https://www.kaggle.com/datasets/mysarahmadbhat/nyc-traffic-accidents

This map was created based on several similar projects.