Interactive Dashboard using TKinter
This is an example of an interactive dashboard using TKinter in order to demonstrate statistics about books ratings and some Market Basket Analysis. 
This dashboard was designed to older adults (65+) focusing on the contrast, as it is known that at this age their eyes get 40% less light, changing the vision. Additionally, titles and labels were set with large font size (12 or more) in order to better visualize the numbers.